Add .unapply and .toOption to Opt[A]
